requirements.txt
azure-functions==1.20.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
phonenumbers==8.13.43
requests==2.32.3
soupsieve==2.6
urllib3==2.2.2
